+++ docs/examples/basic_print_code.py
@@ -1,0 +1,11 @@
+import inspect
+
+import pyreball as pb
+
+
+def add(x, y):
+    return x + y
+
+
+pb.print("An example of a source code printed in HTML:")
+pb.print_source_code(inspect.getsource(add))