+++ docs/examples/basic_print_multi.py
@@ -1,0 +1,3 @@
+import pyreball as pb
+
+pb.print("<div>", "What is 1 + 1?", "<br>", 2, "</div>")