+++ docs/examples/basic_print_title_with_headings.py
@@ -1,0 +1,6 @@
+import pyreball as pb
+
+pb.set_title("Another Report")
+pb.print_h1("Introduction")
+pb.print_h1("Methods")
+pb.print_h1("Conclusion")