+++ docs/examples/custom_tags.py
@@ -2,8 +2,8 @@ import pyreball as pb
 
 pb.set_title("Custom Tags")
 
-pb.print_html(pb.tag("Text displayed in a fixed-width font.", name="pre"))
+pb.print(pb.tag("Text displayed in a fixed-width font.", name="pre"))
 
 hr_tag = pb.tag(name="hr")
 
-pb.print_html(hr_tag)
+pb.print(hr_tag)