+++ pyreball/_common.py
@@ -1,0 +1,4 @@
+from typing import Dict, List, Optional, Union
+
+ClassConfig = Optional[Union[str, List[str]]]
+AttrsConfig = Optional[Dict[str, Optional[str]]]