+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Updated to newer versions of optional dependencies.
 - Added new text elements `div` and `span`. Added new parameters to HTML elements - in particular `cl` and `attrs`.
+- Deprecated `print_html` function. New `print` should be used instead.
+- Deprecated `print_code` function. New `print_source_code` should be used instead.
+- Replaced code-prettify with highlight.js for code blocks.
 
 ## 0.1.1 (2021-09-14)